+++ pkg/integrations/dnsmasq_exporter/dnsmasq_exporter.go
@@ -36,6 +36,11 @@ func (c *Config) CommonConfig() config.Common {
 	return c.Common
 }
 
+// InstanceKey returns the address of the dnsmasq server.
+func (c *Config) InstanceKey(agentKey string) (string, error) {
+	return c.DnsmasqAddress, nil
+}
+
 // NewIntegration converts this config into an instance of an integration.
 func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) {
 	return New(l, c)