+++ pkg/integrations/postgres_exporter/postgres_exporter.go
@@ -9,6 +9,7 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/grafana/agent/pkg/integrations"
 	"github.com/grafana/agent/pkg/integrations/config"
+	"github.com/lib/pq"
 	"github.com/prometheus-community/postgres_exporter/exporter"
 )
 
@@ -43,6 +44,71 @@ func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) 
 	return New(l, c)
 }
 
+// InstanceKey returns a simplified DSN of the first postgresql DSN, or an error if
+// not exactly one DSN is provided.
+func (c *Config) InstanceKey(agentKey string) (string, error) {
+	dsn, err := c.getDataSourceNames()
+	if err != nil {
+		return "", err
+	}
+	if len(dsn) != 1 {
+		return "", fmt.Errorf("can't automatically determine a value for `instance` with %d DSN. either use 1 DSN or manually assign a value for `instance` in the integration config", len(dsn))
+	}
+
+	s, err := parsePostgresURL(dsn[0])
+	if err != nil {
+		return "", fmt.Errorf("cannot parse DSN: %w", err)
+	}
+
+	// Assign default values to s.
+	//
+	// PostgreSQL hostspecs can contain multiple host pairs. We'll assign a host
+	// and port by default, but otherwise just use the hostname.
+	if _, ok := s["host"]; !ok {
+		s["host"] = "localhost"
+		s["port"] = "5432"
+	}
+
+	hostport := s["host"]
+	if p, ok := s["port"]; ok {
+		hostport += fmt.Sprintf(":%s", p)
+	}
+	return fmt.Sprintf("postgresql://%s/%s", hostport, s["dbname"]), nil
+}
+
+func parsePostgresURL(url string) (map[string]string, error) {
+	raw, err := pq.ParseURL(url)
+	if err != nil {
+		return nil, err
+	}
+
+	res := map[string]string{}
+
+	for _, keypair := range strings.Split(raw, " ") {
+		parts := strings.SplitN(keypair, "=", 1)
+		if len(parts) != 2 {
+			panic(fmt.Sprintf("unexpected keypair %s from pq", keypair))
+		}
+
+		res[parts[0]] = parts[1]
+	}
+
+	return res, nil
+}
+
+// getDataSourceNames loads data source names from the config or from the
+// environment, if set.
+func (c *Config) getDataSourceNames() ([]string, error) {
+	dsn := c.DataSourceNames
+	if len(dsn) == 0 {
+		dsn = strings.Split(os.Getenv("POSTGRES_EXPORTER_DATA_SOURCE_NAME"), ",")
+	}
+	if len(dsn) == 0 {
+		return nil, fmt.Errorf("cannot create postgres_exporter; neither postgres_exporter.data_source_name or $POSTGRES_EXPORTER_DATA_SOURCE_NAME is set")
+	}
+	return dsn, nil
+}
+
 func init() {
 	integrations.RegisterIntegration(&Config{})
 }
@@ -50,12 +116,9 @@ func init() {
 // New creates a new postgres_exporter integration. The integration scrapes
 // metrics from a postgres process.
 func New(log log.Logger, c *Config) (integrations.Integration, error) {
-	dsn := c.DataSourceNames
-	if len(dsn) == 0 {
-		dsn = strings.Split(os.Getenv("POSTGRES_EXPORTER_DATA_SOURCE_NAME"), ",")
-	}
-	if len(dsn) == 0 {
-		return nil, fmt.Errorf("cannot create postgres_exporter; neither postgres_exporter.data_source_name or $POSTGRES_EXPORTER_DATA_SOURCE_NAME is set")
+	dsn, err := c.getDataSourceNames()
+	if err != nil {
+		return nil, err
 	}
 
 	e := exporter.NewExporter(